Clean up finite_sum whitespace

+++ abelfunctions/riemanntheta/finite_sum.h
@@ -48,7 +48,7 @@ extern "C" {
 void
 finite_sum_without_derivatives(double* fsum_real, double* fsum_imag,
 			       double* X, double* Yinv, double* T,
-			       double* x, double* y, double* S,
+			       double* x, double* y, int* S,
 			       int g, int N);
 
 
@@ -81,7 +81,7 @@ finite_sum_without_derivatives(double* fsum_real, double* fsum_imag,
 void
 finite_sum_with_derivatives(double* fsum_real, double* fsum_imag,
 			    double* X, double* Yinv, double* T,
-			    double* x, double* y, double* S,
+			    double* x, double* y, int* S,
 			    double* deriv_real, double* deriv_imag, int nderivs,
 			    int g, int N);
